TFP-5919 Legg til saksnummer i kontrakter mot risk og fordel
