+++ vl-kontrakt-fp-formidling/src/main/java/no/nav/foreldrepenger/kontrakter/formidling/kodeverk/Saksnummer.java
@@ -1,0 +1,6 @@
+package no.nav.foreldrepenger.kontrakter.formidling.kodeverk;
+
+import jakarta.validation.constraints.Pattern;
+import jakarta.validation.constraints.Size;
+
+public record Saksnummer(@Size(max = 20) @Pattern(regexp = "^[a-zA-Z0-9_\\-]*$") String saksnummer) { }
