+++ vl-kontrakt-fp-formidling/src/main/java/no/nav/foreldrepenger/kontrakter/formidling/v3/DokumentForhåndsvisDto.java
@@ -4,6 +4,7 @@ import jakarta.validation.Valid;
 import jakarta.validation.constraints.NotNull;
 import no.nav.foreldrepenger.kontrakter.formidling.kodeverk.DokumentMal;
 import no.nav.foreldrepenger.kontrakter.formidling.kodeverk.RevurderingÅrsak;
+import no.nav.foreldrepenger.kontrakter.formidling.kodeverk.Saksnummer;
 
 import java.util.UUID;
 
@@ -18,6 +19,7 @@ import java.util.UUID;
  * @param tittel tittel i FRITEKSTBREV dokumentet. Kun om FRITEKSTBREV mal brukes.
  */
 public record DokumentForhåndsvisDto(@NotNull @Valid UUID behandlingUuid,
+                                     @Valid Saksnummer saksnummer,
                                      @NotNull @Valid DokumentMal dokumentMal,
                                      @Valid RevurderingÅrsak revurderingÅrsak,
                                      String tittel,
