+++ vl-kontrakt-fp-risk/src/main/java/no/nav/foreldrepenger/kontrakter/risk/v1/HentRisikovurderingDto.java
@@ -1,8 +1,11 @@
 package no.nav.foreldrepenger.kontrakter.risk.v1;
 
+import jakarta.validation.Valid;
+import jakarta.validation.constraints.NotNull;
+import no.nav.foreldrepenger.kontrakter.risk.kodeverk.Saksnummer;
+
 import java.util.UUID;
 
-import jakarta.validation.constraints.NotNull;
-
-public record HentRisikovurderingDto(@NotNull UUID konsumentId) {
+public record HentRisikovurderingDto(@NotNull UUID konsumentId,
+                                     @Valid Saksnummer saksnummer) {
 }
